requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2

scikit-learn~=1.5.2